Disables directory removal/creation if specified

+++ build.py
@@ -93,7 +93,6 @@ def main():
     p.add_argument(
         '--root',
         metavar='PATH',
-        default='dist',
         help='directory to store build outputs')
 
     # Parse arguments.
@@ -124,10 +123,13 @@ def main():
     # Create output directory.
     dest = args.root
 
-    if os.path.exists(dest):
-        shutil.rmtree(dest)
+    if dest is None:
+        dest = 'dist'
 
-    os.mkdir(dest)
+        if os.path.exists(dest):
+            shutil.rmtree(dest)
+
+        os.mkdir(dest)
 
     # Export artifacts.
     s = platform.system()
